fixed paralle implemnetation in diff tool
