+++ watertap/tools/parameter_sweep/parameter_sweep_differential.py
@@ -17,16 +17,23 @@ from watertap.tools.parameter_sweep.parameter_sweep import (
     _ParameterSweepBase,
     ParameterSweep,
     return_none,
+    do_execute,
+    return_none,
 )
 from watertap.tools.parallel.single_process_parallel_manager import (
     SingleProcessParallelManager,
 )
 
-
 from pyomo.common.deprecation import deprecation_warning
+from watertap.tools.parameter_sweep.paramter_sweep_parallel_utils import (
+    _ParameterSweepParallelUtils,
+    do_build,
+    do_execute,
+    return_none,
+)
 
 
-class DifferentialParameterSweep(_ParameterSweepBase):
+class DifferentialParameterSweep(_ParameterSweepBase, _ParameterSweepParallelUtils):
     CONFIG = _ParameterSweepBase.CONFIG()
 
     CONFIG.declare(
@@ -86,6 +93,53 @@ class DifferentialParameterSweep(_ParameterSweepBase):
             """,
         ),
     )
+    CONFIG.declare(
+        "build_differential_sweep_specs",
+        ConfigValue(
+            default=None,
+            # domain=function,
+            description="Function for building the differential_sweep_specs",
+            doc="""
+            Must build a specification dictionary that contains details for how to construct the parameter sweep dictionary for differential sweep.
+            This is a nested dictionary where the first level denotes the variable names for which the differential sweep needs to be carried out.
+            The second level denotes various options to be used for wach variable.
+            The number of samples for each differential sweep is specified while initializing the DifferentialParameterSweep object wsing the keyword `num_diff_samples`
+            e.g.
+            
+            {
+                "fs.a": {
+                    "diff_mode": "sum",
+                    "diff_sample_type": NormalSample,
+                    "std_dev": 0.01,
+                    "pyomo_object": m.fs.input["a"],
+                },
+                "fs.b": {
+                    "diff_mode": "product",
+                    "diff_sample_type": UniformSample,
+                    "relative_lb": 0.01,
+                    "relative_ub": 0.01,
+                    "pyomo_object": m.fs.input["b"],
+                },
+                "fs.c": {
+                    "diff_mode": "sum",
+                    "diff_sample_type": GeomSample,
+                    "relative_lb": 0.01,
+                    "relative_ub": 10.0,
+                    "pyomo_object": m.fs.input["c"],
+                },
+            }
+
+            """,
+        ),
+    )
+    CONFIG.declare(
+        "build_differential_sweep_specs_kwargs",
+        ConfigValue(
+            default=dict(),
+            domain=dict,
+            description="Keyword argument for the building differential sweep function",
+        ),
+    )
 
     def __init__(
         self,
@@ -98,7 +152,14 @@ class DifferentialParameterSweep(_ParameterSweepBase):
             raise NotImplementedError
 
     def _create_differential_sweep_params(self, local_values):
-        differential_sweep_specs = self.config.differential_sweep_specs
+        if self.config.differential_sweep_specs is None:
+            # TODO: this should be removd once we get rid of differential_sweep_specs
+            differential_sweep_specs = self.config.differential_sweep_specs
+        else:
+            differential_sweep_specs = self.config.build_differential_sweep_specs(
+                self.model_manager.model,
+                **self.config.build_differential_sweep_specs_kwargs,
+            )
 
         diff_sweep_param = {}
         for ctr, (param, specs) in enumerate(differential_sweep_specs.items()):
@@ -132,8 +193,10 @@ class DifferentialParameterSweep(_ParameterSweepBase):
 
         return diff_sweep_param
 
-    def _check_differential_sweep_key_validity(self, sweep_params):
-        diff_specs_keys = list(self.config.differential_sweep_specs.keys())
+    def _check_differential_sweep_key_validity(
+        self, differential_sweep_spec, sweep_params
+    ):
+        diff_specs_keys = list(differential_sweep_spec.keys())
         sweep_param_keys = list(sweep_params.keys())
 
         if all(key in sweep_param_keys for key in diff_specs_keys):
@@ -145,22 +208,23 @@ class DifferentialParameterSweep(_ParameterSweepBase):
                 "differential_sweep_specs keys don't match with sweep_param keys"
             )
 
-    def _define_differential_sweep_outputs(self, sweep_params):
-        self.differential_outputs = self.outputs
-        if self.outputs is not None:
+    def _define_differential_sweep_outputs(self, outputs, sweep_params):
+        # Currently used in do_build function only (check paramter_sweep_parallel_utils.py)
+        self.differential_outputs = outputs
+        if outputs is not None:
             for key in sweep_params.keys():
                 if key not in self.config.differential_sweep_specs.keys():
                     self.differential_outputs[key] = sweep_params[key].pyomo_object
 
-    def _create_local_output_skeleton(self, model, sweep_params, outputs, num_samples):
-        output_dict = super()._create_local_output_skeleton(
-            model, sweep_params, outputs, num_samples
-        )
-        output_dict["nominal_idx"] = np.arange(
-            num_samples, dtype=float
-        )  # [*range(num_samples)]
-        output_dict["differential_idx"] = np.array([np.nan] * num_samples)
-        return output_dict
+    # def _create_local_output_skeleton(self, model, sweep_params, outputs, num_samples):
+    #     output_dict = super()._create_local_output_skeleton(
+    #         model, sweep_params, outputs, num_samples
+    #     )
+    #     output_dict["nominal_idx"] = np.arange(
+    #         num_samples, dtype=float
+    #     )  # [*range(num_samples)]
+    #     output_dict["differential_idx"] = np.array([np.nan] * num_samples)
+    #     return output_dict
 
     def _append_differential_results(self, local_output_dict, diff_results_dict):
         for idx, diff_sol in diff_results_dict.items():
@@ -210,109 +274,57 @@ class DifferentialParameterSweep(_ParameterSweepBase):
                                 )
                             )
 
-    def _collect_local_inputs(self, local_results_dict):
-        num_local_samples = len(local_results_dict["solve_successful"])
-        local_inputs = np.zeros(
-            (num_local_samples, len(local_results_dict["sweep_params"])),
-            dtype=float,
-        )
+    # def _create_global_output(self, local_output_dict):  # , req_num_samples=None):
+    #     global_output_dict = super()._create_global_output(local_output_dict)
 
-        for i, (key, item) in enumerate(local_results_dict["sweep_params"].items()):
-            local_inputs[:, i] = item["value"]
+    #     # We now need to get the mapping array. This only needs to happen on root
+    #     local_num_cases_all = len(local_output_dict["solve_successful"])
+    #     # AllGather the total size of the value array on each MPI rank
+    #     sample_split_arr = self.parallel_manager.combine_data_with_peers(
+    #         local_num_cases_all
+    #     )
+    #     num_total_samples = sum(sample_split_arr)
 
-        return local_inputs
+    #     # AllGather nominal values for creating the parallel offset
+    #     nominal_sample_split_arr = self.parallel_manager.combine_data_with_peers(
+    #         self.n_nominal_local
+    #     )
 
-    def _aggregate_input_arr(self, global_results_dict, num_global_samples):
-        global_values = np.zeros(
-            (num_global_samples, len(global_results_dict["sweep_params"])),
-            dtype=float,
-        )
+    #     # We need to create a global index and offset items accordingly. This
+    #     # needs to happen on all ranks/workers.
+    #     my_rank = self.parallel_manager.get_rank()
+    #     offset = 0
+    #     if my_rank > 0:
+    #         offset = sum(nominal_sample_split_arr[:my_rank])
+    #     local_output_dict["nominal_idx"] = local_output_dict["nominal_idx"] + offset
+    #     local_output_dict["differential_idx"] = (
+    #         local_output_dict["differential_idx"] + offset
+    #     )
 
-        if self.parallel_manager.is_root_process():
-            for i, (key, item) in enumerate(
-                global_results_dict["sweep_params"].items()
-            ):
-                global_values[:, i] = item["value"]
+    #     # Resize global index array
+    #     if self.parallel_manager.is_root_process():
+    #         global_output_dict["nominal_idx"] = np.zeros(num_total_samples, dtype=float)
+    #         global_output_dict["differential_idx"] = np.zeros(
+    #             num_total_samples, dtype=float
+    #         )
 
-        self.parallel_manager.sync_array_with_peers(global_values)
+    #     # Now we need to collect it on global_output_dict
+    #     self.parallel_manager.gather_arrays_to_root(
+    #         sendbuf=local_output_dict["nominal_idx"],
+    #         recvbuf_spec=(
+    #             global_output_dict["nominal_idx"],
+    #             sample_split_arr,
+    #         ),
+    #     )
+    #     self.parallel_manager.gather_arrays_to_root(
+    #         sendbuf=local_output_dict["differential_idx"],
+    #         recvbuf_spec=(
+    #             global_output_dict["differential_idx"],
+    #             sample_split_arr,
+    #         ),
+    #     )
 
-        return global_values
-
-    def _aggregate_results(self, local_output_dict):
-        # Create the global results dictionary
-        global_results_dict = self._create_global_output(local_output_dict)
-
-        # Broadcast the number of global samples to all ranks
-        num_global_samples = len(global_results_dict["solve_successful"])
-        num_global_samples = self.parallel_manager.sync_pyobject_with_peers(
-            num_global_samples
-        )
-
-        global_results_arr = self._aggregate_results_arr(
-            global_results_dict, num_global_samples
-        )
-        global_input_values = self._aggregate_input_arr(
-            global_results_dict, num_global_samples
-        )
-
-        return (
-            global_results_dict,
-            global_results_arr,
-            global_input_values,
-            num_global_samples,
-        )
-
-    def _create_global_output(self, local_output_dict):  # , req_num_samples=None):
-        global_output_dict = super()._create_global_output(local_output_dict)
-
-        # We now need to get the mapping array. This only needs to happen on root
-        local_num_cases_all = len(local_output_dict["solve_successful"])
-        # AllGather the total size of the value array on each MPI rank
-        sample_split_arr = self.parallel_manager.combine_data_with_peers(
-            local_num_cases_all
-        )
-        num_total_samples = sum(sample_split_arr)
-
-        # AllGather nominal values for creating the parallel offset
-        nominal_sample_split_arr = self.parallel_manager.combine_data_with_peers(
-            self.n_nominal_local
-        )
-
-        # We need to create a global index and offset items accordingly. This
-        # needs to happen on all ranks/workers.
-        my_rank = self.parallel_manager.get_rank()
-        offset = 0
-        if my_rank > 0:
-            offset = sum(nominal_sample_split_arr[:my_rank])
-        local_output_dict["nominal_idx"] = local_output_dict["nominal_idx"] + offset
-        local_output_dict["differential_idx"] = (
-            local_output_dict["differential_idx"] + offset
-        )
-
-        # Resize global index array
-        if self.parallel_manager.is_root_process():
-            global_output_dict["nominal_idx"] = np.zeros(num_total_samples, dtype=float)
-            global_output_dict["differential_idx"] = np.zeros(
-                num_total_samples, dtype=float
-            )
-
-        # Now we need to collect it on global_output_dict
-        self.parallel_manager.gather_arrays_to_root(
-            sendbuf=local_output_dict["nominal_idx"],
-            recvbuf_spec=(
-                global_output_dict["nominal_idx"],
-                sample_split_arr,
-            ),
-        )
-        self.parallel_manager.gather_arrays_to_root(
-            sendbuf=local_output_dict["differential_idx"],
-            recvbuf_spec=(
-                global_output_dict["differential_idx"],
-                sample_split_arr,
-            ),
-        )
-
-        return global_output_dict
+    #     return global_output_dict
 
     def _run_differential_sweep(self, local_value):
         diff_sweep_param_dict = self._create_differential_sweep_params(local_value)
@@ -329,7 +341,11 @@ class DifferentialParameterSweep(_ParameterSweepBase):
         # pass model_manager from refernce sweep, to diff sweep
         # so we don't have to reijnit he model
         diff_ps.model_manager = self.model_manager
-
+        print(
+            "diff model passed state",
+            self.model_manager.is_initialized,
+            diff_sweep_param_dict,
+        )
         _, differential_sweep_output_dict = diff_ps.parameter_sweep(
             diff_ps.model_manager.model,
             diff_sweep_param_dict,
@@ -424,6 +440,7 @@ class DifferentialParameterSweep(_ParameterSweepBase):
                                 and will not work with future implementations of parallelism.",
                 version="0.10.0",
             )
+
         # This should be depreciated in future versions
         self.config.build_model = build_model
         self.config.build_sweep_params = build_sweep_params
@@ -435,72 +452,89 @@ class DifferentialParameterSweep(_ParameterSweepBase):
         model = build_model(**build_model_kwargs)
         sweep_params = build_sweep_params(model, **build_sweep_params_kwargs)
         sweep_params, sampling_type = self._process_sweep_params(sweep_params)
-
-        # Check if the keys in the differential sweep specs exist in sweep params
-        self._check_differential_sweep_key_validity(sweep_params)
-
-        # Define differential sweep outputs
-        self.outputs = self.config.build_outputs(
-            model, *self.config.build_outputs_kwargs
+        differential_sweep_spec = self.config.build_differential_sweep_specs(
+            model, **self.config.build_differential_sweep_specs_kwargs
         )
-        self._define_differential_sweep_outputs(sweep_params)
+        # Check if the keys in the differential sweep specs exist in sweep params
+        self._check_differential_sweep_key_validity(
+            differential_sweep_spec, sweep_params
+        )
 
         # Set the seed before sampling
         self.seed = seed
         np.random.seed(self.seed)
 
         # Enumerate/Sample the parameter space
-        global_values = self._build_combinations(
+        all_parameter_combinations = self._build_combinations(
             sweep_params, sampling_type, num_samples
         )
-
-        # divide the workload between processors
-        local_values = self._divide_combinations(global_values)
-        self.n_nominal_local = np.shape(local_values)[0]
-
-        # Check if the outputs have the name attribute. If not, assign one.
-        if self.outputs is not None:
-            self.assign_variable_names(model, self.outputs)
-
-        # Create a dictionary to store all the differential ps_objects
-        self.diff_ps_dict = {}
-
-        # Do the Loop
-        if self.config.custom_do_param_sweep is None:
-            local_results_dict = self._do_param_sweep(
-                sweep_params,
-                self.outputs,
-                local_values,
-            )
-        else:
-            local_results_dict = self.config.custom_do_param_sweep(
-                self,
-                sweep_params,
-                self.outputs,
-                local_values,
-                **self.config.custom_do_param_sweep_kwargs,
-            )
-
-        # re-writing local_values
-        local_values = self._collect_local_inputs(local_results_dict)
-
-        # Aggregate results on Master
-        (
-            global_results_dict,
-            global_results_arr,
-            global_input_arr,
-            num_global_samples,
-        ) = self._aggregate_results(local_results_dict)
-
-        # Save to file
-        global_save_data = self.writer.save_results(
-            sweep_params,
-            local_values,
-            global_input_arr,
-            local_results_dict,
-            global_results_dict,
-            global_results_arr,
-            self.parallel_manager.get_rank(),
+        all_results = self.run_scatter_gather(
+            all_parameter_combinations, DifferentialParameterSweep
         )
+        global_sweep_results_dict = self._combine_gather_results(all_results)
+        combined_output_arr = self._combine_output_array(global_sweep_results_dict)
+        # save the results for all simulations run by this process and its children
+        for results in self.parallel_manager.results_from_local_tree(all_results):
+            self.writer.save_results(
+                sweep_params,
+                results.parameters,
+                all_parameter_combinations,
+                results.results,
+                global_sweep_results_dict,
+                combined_output_arr,
+                process_number=results.process_number,
+            )
 
-        return global_save_data, global_results_dict
+        global_save_data = np.hstack((all_parameter_combinations, combined_output_arr))
+        return global_save_data, global_sweep_results_dict
+
+    # # divide the workload between processors
+    # local_values = self._divide_combinations(all_parameter_combinations)
+    # self.n_nominal_local = np.shape(local_values)[0]
+
+    # # Check if the outputs have the name attribute. If not, assign one.
+    # if self.outputs is not None:
+    #     self.assign_variable_names(model, self.outputs)
+
+    # # Create a dictionary to store all the differential ps_objects
+    # self.diff_ps_dict = {}
+
+    # # Do the Loop
+    # if self.config.custom_do_param_sweep is None:
+    #     local_results_dict = self._do_param_sweep(
+    #         sweep_params,
+    #         self.outputs,
+    #         local_values,
+    #     )
+    # else:
+    #     local_results_dict = self.config.custom_do_param_sweep(
+    #         self,
+    #         sweep_params,
+    #         self.outputs,
+    #         local_values,
+    #         **self.config.custom_do_param_sweep_kwargs,
+    #     )
+
+    # # re-writing local_values
+    # local_values = self._collect_local_inputs(local_results_dict)
+
+    # # Aggregate results on Master
+    # (
+    #     global_results_dict,
+    #     global_results_arr,
+    #     global_input_arr,
+    #     num_global_samples,
+    # ) = self._aggregate_results(local_results_dict)
+
+    # # Save to file
+    # global_save_data = self.writer.save_results(
+    #     sweep_params,
+    #     local_values,
+    #     global_input_arr,
+    #     local_results_dict,
+    #     global_results_dict,
+    #     global_results_arr,
+    #     self.parallel_manager.get_rank(),
+    # )
+
+    # return global_save_data, global_results_dict
